requirements.txt
Flask==2.1.3
google-cloud-storage==2.14.0